+++ test/weighted_directed_graph_test.dart
@@ -108,6 +108,15 @@ void main() {
       graph.clear();
       expect(graph.isEmpty, true);
     });
+    test('clearEdges', () {
+      final graph = WeightedDirectedGraph.of(graph0);
+      expect(graph.sortedVertices, graph0.sortedVertices);
+      graph.clearEdges();
+      expect(graph.sortedVertices, graph0.sortedVertices);
+      for (var vertex in graph.sortedVertices) {
+        expect(graph.edges(vertex), isEmpty);
+      }
+    });
   });
   group('Graph data:', () {
     test('edges().', () {